requirements.txt
rpi-gpio
seeed_dht